requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
frozenlist==1.3.3
getuseragent==0.0.7
idna==3.4
iniconfig==2.0.0
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pycountry==22.3.5
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==1.26.15
yarl==1.8.2
